+++ CITreeViewClasses/CITreeView.swift
@@ -1,30 +1,29 @@
 //
-//  TreeView.swift
-//  TreeViewExample
+//  CITreeViewController.swift
+//  CITreeView
 //
-//  Created by Cenk IŞIK on 11.01.2018.
-//  Copyright © 2018 Cenk IŞIK. All rights reserved.
+//  Created by Apple on 24.01.2018.
+//  Copyright © 2018 Cenk Işık. All rights reserved.
 //
 
 import UIKit
 
 @objc
-public protocol CITreeViewDataSource : NSObjectProtocol{
-    func treeView(_ treeView:CITreeView, atIndexPath indexPath:IndexPath, withTreeViewNode treeViewNode:CITreeViewNode) -> UITableViewCell
-    func treeViewSelectedNodeChildren(for treeViewNodeItem:AnyObject) -> [AnyObject]
-    func treeViewDataArray() -> [AnyObject]
+public protocol CITreeViewDataSource: NSObjectProtocol {
+    func treeView(_ treeView: CITreeView, cellForRowAt indexPath: IndexPath, with treeViewNode: CITreeViewNode) -> UITableViewCell
+    func treeViewSelectedNodeChildren(for treeViewNodeItem: Any) -> [Any]
+    func treeViewDataArray() -> [Any]
 }
 
 @objc
-public protocol CITreeViewDelegate : NSObjectProtocol{
-    
-    func treeView(_ treeView: CITreeView, heightForRowAt indexPath: IndexPath, withTreeViewNode treeViewNode:CITreeViewNode) -> CGFloat
-    func treeView(_ treeView: CITreeView, didSelectRowAt treeViewNode:CITreeViewNode, atIndexPath indexPath:IndexPath)
-    func treeView(_ treeView: CITreeView, didDeselectRowAt treeViewNode:CITreeViewNode, atIndexPath indexPath: IndexPath)
-    func willExpandTreeViewNode(treeViewNode:CITreeViewNode, atIndexPath: IndexPath)
-    func didExpandTreeViewNode(treeViewNode:CITreeViewNode, atIndexPath: IndexPath)
-    func willCollapseTreeViewNode(treeViewNode:CITreeViewNode, atIndexPath: IndexPath)
-    func didCollapseTreeViewNode(treeViewNode:CITreeViewNode, atIndexPath: IndexPath)
+public protocol CITreeViewDelegate: NSObjectProtocol {
+    func treeView(_ treeView: CITreeView, heightForRowAt indexPath: IndexPath, with treeViewNode: CITreeViewNode) -> CGFloat
+    func treeView(_ treeView: CITreeView, didSelectRowAt treeViewNode: CITreeViewNode, at indexPath: IndexPath)
+    func treeView(_ treeView: CITreeView, didDeselectRowAt treeViewNode: CITreeViewNode, at indexPath: IndexPath)
+    func treeViewNode(_ treeViewNode: CITreeViewNode, willExpandAt indexPath: IndexPath)
+    func treeViewNode(_ treeViewNode: CITreeViewNode, didExpandAt indexPath: IndexPath)
+    func treeViewNode(_ treeViewNode: CITreeViewNode, willCollapseAt indexPath: IndexPath)
+    func treeViewNode(_ treeViewNode: CITreeViewNode, didCollapseAt indexPath: IndexPath)
     
 }
 
@@ -79,7 +78,7 @@ public class CITreeView: UITableView {
             return
         }
         
-        if (treeViewDataSource.treeViewDataArray()).count > treeViewController.treeViewNodes.count {
+        if treeViewDataSource.treeViewDataArray().count > treeViewController.treeViewNodes.count {
             mainDataArray = [CITreeViewNode]()
             treeViewController.treeViewNodes.removeAll()
             for item in treeViewDataSource.treeViewDataArray() {
@@ -118,13 +117,13 @@ public class CITreeView: UITableView {
             self.performBatchUpdates({
                 deleteRows()
             }, completion: { (complete) in
-                treeViewDelegate.didCollapseTreeViewNode(treeViewNode: treeViewNode, atIndexPath:indexPath)
+                treeViewDelegate.treeViewNode(treeViewNode, didCollapseAt: indexPath)
                 completion()
             })
         } else {
             CATransaction.begin()
             CATransaction.setCompletionBlock({
-                treeViewDelegate.didCollapseTreeViewNode(treeViewNode: treeViewNode, atIndexPath: indexPath)
+                treeViewDelegate.treeViewNode(treeViewNode, didCollapseAt: indexPath)
                 completion()
             })
             deleteRows()
@@ -138,12 +137,12 @@ public class CITreeView: UITableView {
             self.performBatchUpdates({
                 insertRows()
             }, completion: { (complete) in
-                treeViewDelegate.didExpandTreeViewNode(treeViewNode: treeViewNode, atIndexPath: indexPath)
+                treeViewDelegate.treeViewNode(treeViewNode, didExpandAt: indexPath)
             })
         } else {
             CATransaction.begin()
             CATransaction.setCompletionBlock({
-                treeViewDelegate.didExpandTreeViewNode(treeViewNode: treeViewNode, atIndexPath: indexPath)
+                treeViewDelegate.treeViewNode(treeViewNode, didExpandAt: indexPath)
             })
             insertRows()
             CATransaction.commit()
@@ -159,10 +158,10 @@ public class CITreeView: UITableView {
         }
         return cells
     }
+    
     public func expandAllRows() {
         treeViewController.expandAllRows()
         reloadDataWithoutChangingRowStates()
-        
     }
     
     public func collapseAllRows() {
@@ -171,11 +170,11 @@ public class CITreeView: UITableView {
     }
 }
 
-extension CITreeView : UITableViewDelegate {
+extension CITreeView: UITableViewDelegate {
     
     public func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         let treeViewNode = treeViewController.getTreeViewNode(atIndex: indexPath.row)
-        return (self.treeViewDelegate?.treeView(tableView as! CITreeView,heightForRowAt: indexPath,withTreeViewNode :treeViewNode))!
+        return (self.treeViewDelegate?.treeView(tableView as! CITreeView, heightForRowAt: indexPath, with: treeViewNode))!
     }
     
     public func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
@@ -183,14 +182,12 @@ extension CITreeView : UITableViewDelegate {
         guard let treeViewDelegate = self.treeViewDelegate else { return }
         
         if let justSelectedTreeViewNode = selectedTreeViewNode {
-            treeViewDelegate.treeView(tableView as! CITreeView, didSelectRowAt: justSelectedTreeViewNode, atIndexPath: indexPath)
+            treeViewDelegate.treeView(tableView as! CITreeView, didSelectRowAt: justSelectedTreeViewNode, at: indexPath)
             var willExpandIndexPath = indexPath
             if justSelectedTreeViewNode.expand {
                 treeViewController.collapseRows(for: justSelectedTreeViewNode, atIndexPath: indexPath)
                 collapseRows(for: justSelectedTreeViewNode, atIndexPath: indexPath){}
-            }
-            else
-            {
+            } else {
                 if collapseNoneSelectedRows,
                     selectedTreeViewNode?.level == 0,
                     let collapsedTreeViewNode = treeViewController.collapseAllRowsExceptOne(),
@@ -205,8 +202,7 @@ extension CITreeView : UITableViewDelegate {
                         self.treeViewController.expandRows(atIndexPath: willExpandIndexPath, with: justSelectedTreeViewNode, openWithChildrens: false)
                         self.expandRows(for: justSelectedTreeViewNode, withSelected: indexPath)
                     }
-                    
-                }else{
+                } else {
                     treeViewController.expandRows(atIndexPath: willExpandIndexPath, with: justSelectedTreeViewNode, openWithChildrens: false)
                     expandRows(for: justSelectedTreeViewNode, withSelected: indexPath)
                 }
@@ -215,27 +211,27 @@ extension CITreeView : UITableViewDelegate {
     }
 }
 
-extension CITreeView : UITableViewDataSource {
+extension CITreeView: UITableViewDataSource {
     public func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return treeViewController.treeViewNodes.count
     }
     
     public func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let treeViewNode = treeViewController.getTreeViewNode(atIndex: indexPath.row)
-        return (self.treeViewDataSource?.treeView(tableView as! CITreeView, atIndexPath: indexPath, withTreeViewNode: treeViewNode))!
+        return (self.treeViewDataSource?.treeView(tableView as! CITreeView, cellForRowAt: indexPath, with: treeViewNode))!
     }
 }
 
-extension CITreeView : CITreeViewControllerDelegate {
-    public func getChildren(forTreeViewNodeItem item: AnyObject, with indexPath: IndexPath) -> [AnyObject] {
-        return (self.treeViewDataSource?.treeViewSelectedNodeChildren(for: item))!
+extension CITreeView: CITreeViewControllerDelegate {
+    public func getChildren(for treeViewNodeItem: Any, at indexPath: IndexPath) -> [Any] {
+        return (self.treeViewDataSource?.treeViewSelectedNodeChildren(for: treeViewNodeItem)) ?? []
     }
     
-    public func willCollapseTreeViewNode(treeViewNode: CITreeViewNode, atIndexPath: IndexPath) {
-        self.treeViewDelegate?.willCollapseTreeViewNode(treeViewNode: treeViewNode, atIndexPath: atIndexPath)
+    public func willCollapseTreeViewNode(_ treeViewNode: CITreeViewNode, at indexPath: IndexPath) {
+        self.treeViewDelegate?.treeViewNode(treeViewNode, willCollapseAt: indexPath)
     }
     
-    public func willExpandTreeViewNode(treeViewNode: CITreeViewNode, atIndexPath: IndexPath) {
-        self.treeViewDelegate?.willExpandTreeViewNode(treeViewNode: treeViewNode, atIndexPath: atIndexPath)
+    public func willExpandTreeViewNode(_ treeViewNode: CITreeViewNode, at indexPath: IndexPath) {
+        self.treeViewDelegate?.treeViewNode(treeViewNode, willExpandAt: indexPath)
     }
 }
