+++ CITreeViewClasses/CITreeViewController.swift
@@ -8,30 +8,29 @@
 
 import Foundation
 
-public protocol CITreeViewControllerDelegate : NSObjectProtocol {
-    func getChildren(forTreeViewNodeItem item:AnyObject, with indexPath:IndexPath) -> [AnyObject]
-    func willExpandTreeViewNode(treeViewNode:CITreeViewNode, atIndexPath: IndexPath)
-    func willCollapseTreeViewNode(treeViewNode:CITreeViewNode, atIndexPath: IndexPath)
+public protocol CITreeViewControllerDelegate: NSObjectProtocol {
+    func getChildren(for treeViewNodeItem: Any, at indexPath: IndexPath) -> [Any]
+    func willExpandTreeViewNode(_ treeViewNode: CITreeViewNode, at indexPath: IndexPath)
+    func willCollapseTreeViewNode(_ treeViewNode: CITreeViewNode, at indexPath: IndexPath)
 }
 
-public class CITreeViewController:NSObject  {
+public class CITreeViewController: NSObject {
     
-    var treeViewNodes:[CITreeViewNode] = []
-    var indexPathsArray:[IndexPath] = []
-    weak var treeViewControllerDelegate:CITreeViewControllerDelegate?
+    var treeViewNodes: [CITreeViewNode] = []
+    var indexPathsArray: [IndexPath] = []
+    weak var treeViewControllerDelegate: CITreeViewControllerDelegate?
     
-    init(treeViewNodes : [CITreeViewNode]) {
+    init(treeViewNodes: [CITreeViewNode]) {
         self.treeViewNodes = treeViewNodes
     }
     
     //MARK: Tree View Nodes Functions
-    func addTreeViewNode(with item:AnyObject){
+    func addTreeViewNode(with item: Any) {
         let treeViewNode = CITreeViewNode(item: item)
         treeViewNodes.append(treeViewNode)
     }
     
-    func getTreeViewNode(atIndex index: Int) -> CITreeViewNode
-    {
+    func getTreeViewNode(atIndex index: Int) -> CITreeViewNode {
         if treeViewNodes.indices.contains(index){
             return treeViewNodes[index]
         }
@@ -39,47 +38,45 @@ public class CITreeViewController:NSObject  {
     }
     
     func index(of treeViewNode: CITreeViewNode) -> Int? {
-        return treeViewNodes.firstIndex(of: treeViewNode)
+        return treeViewNodes.index(of: treeViewNode)
     }
     
-    func insertTreeViewNode(parent parentTreeViewNode:CITreeViewNode, with item:AnyObject, to index : Int)
-    {
+    func insertTreeViewNode(parent parentTreeViewNode: CITreeViewNode, with item: Any, to index: Int) {
         let treeViewNode = CITreeViewNode(item: item)
         treeViewNode.parentNode = parentTreeViewNode
         treeViewNodes.insert(treeViewNode, at: index)
     }
     
-    func removeTreeViewNodesAtRange(from start:Int , to end:Int)
-    {
+    func removeTreeViewNodesAtRange(from start:Int , to end:Int) {
         treeViewNodes.removeSubrange(start ... end)
     }
     
-    func setExpandTreeViewNode(atIndex index:Int){
+    func setExpandTreeViewNode(atIndex index:Int) {
         treeViewNodes[index].expand = true
     }
     
-    func setCollapseTreeViewNode(atIndex index:Int){
+    func setCollapseTreeViewNode(atIndex index:Int) {
         treeViewNodes[index].expand = false
     }
     
-    func setLevelTreeViewNode(atIndex index:Int, to level:Int){
+    func setLevelTreeViewNode(atIndex index: Int, to level: Int) {
         treeViewNodes[index].level = level + 1
     }
     
     // MARK: Expand Rows
     
-    func addIndexPath(withRow row:Int){
-        let indexPath = IndexPath(row: row , section: 0)
+    func addIndexPath(withRow row: Int) {
+        let indexPath = IndexPath(row: row, section: 0)
         indexPathsArray.append(indexPath)
     }
     
-    func expandRows(atIndexPath indexPath:IndexPath, with selectedTreeViewNode:CITreeViewNode){
-        let children = self.treeViewControllerDelegate?.getChildren(forTreeViewNodeItem: selectedTreeViewNode.item, with: indexPath)
+    func expandRows(atIndexPath indexPath: IndexPath, with selectedTreeViewNode: CITreeViewNode) {
+        let children = self.treeViewControllerDelegate?.getChildren(for: selectedTreeViewNode.item, at: indexPath)
         indexPathsArray = [IndexPath]()
         var row = indexPath.row + 1
         
         if (children?.count)! > 0 {
-            self.treeViewControllerDelegate?.willExpandTreeViewNode(treeViewNode: selectedTreeViewNode, atIndexPath: indexPath)
+            self.treeViewControllerDelegate?.willExpandTreeViewNode(selectedTreeViewNode, at: indexPath)
             setExpandTreeViewNode(atIndex: indexPath.row)
         }
         
@@ -92,11 +89,11 @@ public class CITreeViewController:NSObject  {
     }
     
     // MARK: Collapse Rows
-    func removeIndexPath(withRow row:inout Int, and indexPath:IndexPath){
+    func removeIndexPath(withRow row: inout Int, and indexPath:IndexPath) {
         let treeViewNode = getTreeViewNode(atIndex: row)
-        let children = self.treeViewControllerDelegate?.getChildren(forTreeViewNodeItem: treeViewNode.item, with: indexPath)
+        let children = self.treeViewControllerDelegate?.getChildren(for: treeViewNode.item, at: indexPath)
         
-        let index = IndexPath(row: row , section: indexPath.section)
+        let index = IndexPath(row: row, section: indexPath.section)
         indexPathsArray.append(index)
         row += 1
         
@@ -107,14 +104,14 @@ public class CITreeViewController:NSObject  {
         }
     }
     
-    func collapseRows(for treeViewNode :CITreeViewNode, atIndexPath indexPath:IndexPath){
-        guard let treeViewControllerDelegate = self.treeViewControllerDelegate else {return}
-        let treeViewNodeChildren = treeViewControllerDelegate.getChildren(forTreeViewNodeItem: treeViewNode.item, with: indexPath)
+    func collapseRows(for treeViewNode: CITreeViewNode, atIndexPath indexPath: IndexPath) {
+        guard let treeViewControllerDelegate = self.treeViewControllerDelegate else { return }
+        let treeViewNodeChildren = treeViewControllerDelegate.getChildren(for: treeViewNode.item, at: indexPath)
         indexPathsArray = [IndexPath]()
         var row = indexPath.row + 1
         
         if treeViewNodeChildren.count > 0 {
-            treeViewControllerDelegate.willCollapseTreeViewNode(treeViewNode: treeViewNode, atIndexPath: indexPath)
+            treeViewControllerDelegate.willCollapseTreeViewNode(treeViewNode, at: indexPath)
         }
         
         setCollapseTreeViewNode(atIndex: indexPath.row)
@@ -122,14 +119,14 @@ public class CITreeViewController:NSObject  {
         for _ in treeViewNodeChildren{
             removeIndexPath(withRow: &row, and: indexPath)
         }
+        
         if indexPathsArray.count > 0 {
             removeTreeViewNodesAtRange(from: (indexPathsArray.first?.row)!, to: (indexPathsArray.last?.row)!)
         }
-        
     }
     
     
-    @discardableResult func collapseAllRowsExceptOne() -> CITreeViewNode?{
+    @discardableResult func collapseAllRowsExceptOne() -> CITreeViewNode? {
         indexPathsArray = [IndexPath]()
         var collapsedTreeViewNode:CITreeViewNode? = nil
         var indexPath = IndexPath(row: 0, section: 0)
@@ -143,16 +140,16 @@ public class CITreeViewController:NSObject  {
         return collapsedTreeViewNode
     }
     @discardableResult
-    func expandRows(atIndexPath indexPath:IndexPath, with selectedTreeViewNode:CITreeViewNode, openWithChildrens:Bool) -> Int{
-        guard let treeViewControllerDelegate = self.treeViewControllerDelegate else {return 0}
-        let treeViewNodeChildren = treeViewControllerDelegate.getChildren(forTreeViewNodeItem: selectedTreeViewNode.item, with: indexPath)
+    func expandRows(atIndexPath indexPath: IndexPath, with selectedTreeViewNode: CITreeViewNode, openWithChildrens: Bool) -> Int {
+        guard let treeViewControllerDelegate = self.treeViewControllerDelegate else { return 0 }
+        let treeViewNodeChildren = treeViewControllerDelegate.getChildren(for: selectedTreeViewNode.item, at: indexPath)
         indexPathsArray = [IndexPath]()
         var row = indexPath.row + 1
         setExpandTreeViewNode(atIndex: indexPath.row)
         
         if treeViewNodeChildren.count > 0 {
-            treeViewControllerDelegate.willExpandTreeViewNode(treeViewNode: selectedTreeViewNode, atIndexPath: indexPath)
-            for item in treeViewNodeChildren{
+            treeViewControllerDelegate.willExpandTreeViewNode(selectedTreeViewNode, at: indexPath)
+            for item in treeViewNodeChildren {
                 addIndexPath(withRow: row)
                 insertTreeViewNode(parent: selectedTreeViewNode, with: item, to: row)
                 setLevelTreeViewNode(atIndex: row, to: selectedTreeViewNode.level)
@@ -160,10 +157,9 @@ public class CITreeViewController:NSObject  {
                     let treeViewNode = getTreeViewNode(atIndex: row)
                     let indexPath = IndexPath(row: row, section: 0)
                     row = expandRows(atIndexPath: indexPath, with: treeViewNode, openWithChildrens: openWithChildrens)
-                }else {
+                } else {
                     row += 1
                 }
-                
             }
         }
         return row
@@ -174,10 +170,10 @@ public class CITreeViewController:NSObject  {
         var indexPath = IndexPath(row: 0, section: 0)
         for treeViewNode in treeViewNodes {
             indexPath = getIndexPathOfTreeViewNode(treeViewNode: treeViewNode)
-            if  treeViewNode.level != 0 {
+            if treeViewNode.level != 0 {
                 setCollapseTreeViewNode(atIndex: indexPath.row)
                 treeViewNodes.remove(at: indexPath.row)
-            }else{
+            } else {
                 setCollapseTreeViewNode(atIndex: indexPath.row)
             }
         }
@@ -195,7 +191,7 @@ public class CITreeViewController:NSObject  {
         }
     }
     
-    func getIndexPathOfTreeViewNode(treeViewNode:CITreeViewNode) -> IndexPath {
+    func getIndexPathOfTreeViewNode(treeViewNode: CITreeViewNode) -> IndexPath {
         for (index,node) in treeViewNodes.enumerated() {
             if treeViewNode == node {
                 return IndexPath(row: index, section: 0)
